pyproject.toml
[project]
name = "umie-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.2.3",
    "qdrant-client>=1.12.1",
    "requests>=2.32.3",
    "jsonlines>=4.0.0",
    "opencv-python>=4.10.0.84",
    "pillow>=11.0.0",
    "litellm>=1.52.3",
    "scikit-learn>=1.5.2",
    "matplotlib>=3.9.2",
    "seaborn>=0.13.2",
    "ipykernel>=6.29.5",
    "typer>=0.13.1",
    "python-dotenv>=1.0.1",
    "click>=8.1.7",
    "alive-progress>=3.2.0",
    "datasets>=3.1.0",
    "json-repair>=0.30.2",
]
